En este Momento funciona el metodo de agregar cartas a otra Tabla, pero aun no se puede hacer el drag and drop de cada carta

+++ Headers/carta.h
@@ -21,7 +21,7 @@ public:
     string srcImg;
     string imgDefault;
     bool isVisible;
-    bool isClick(sf::Vector2i vec);
+    bool isClick(sf::Vector2f vec);
     sf::Texture texture;
     void setPos(int x, int y);
     void show();
